+++ earthlib/geelib/BrightMask.py
