+++ earthlib/geelib/ShadeMask.py
