+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "earthlib"
-version = "0.3.2"
+version = "0.4.0"
 description = "A global spectral library with earth engine tools for satellite land cover mapping."
 authors = ["earth-chris <cbanders@alumni.stanford.edu>"]
 license = "MIT"
@@ -8,11 +8,13 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.9"
-earthengine-api = ">=0.1.317"
 numpy = ">=1.21.5"
 pandas = ">=1.3.5"
 spectral = ">=0.22.4"
 tqdm = ">=4.63.0"
+
+[tool.poetry.extras]
+ee = ["earthengine-api>=0.1.317"]
 
 [tool.poetry.group.dev.dependencies]
 ipython = "^8.5.0"
